Organize functions into controller folder
